+++ etna/models/nn/tft.py
@@ -53,7 +53,6 @@ class TFTModel(_DeepCopyMixin, SaveNNMixin, PredictionIntervalContextIgnorantAbs
         loss: "MultiHorizonMetric" = None,
         trainer_kwargs: Optional[Dict[str, Any]] = None,
         quantiles_kwargs: Optional[Dict[str, Any]] = None,
-        *args,
         **kwargs,
     ):
         """
@@ -113,6 +112,7 @@ class TFTModel(_DeepCopyMixin, SaveNNMixin, PredictionIntervalContextIgnorantAbs
         self.trainer: Optional[pl.Trainer] = None
         self._last_train_timestamp = None
         self._freq: Optional[str] = None
+        self.kwargs = kwargs
 
     def _from_dataset(self, ts_dataset: TimeSeriesDataSet) -> LightningModule:
         """